pyproject.toml
[tool.poetry]
name = "tetrisai"
version = "0.1.0"
description = ""
authors = ["Chris Vaughn <chris@cvaughn.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.20.3"
opencv-python = "^4.5.2"
pyobjc-framework-Quartz = "^7.2"
mss = "^6.1.0"
pytesseract = "^0.3.7"
progressbar = "^2.5"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
